fix: final set of build error fixes
