+++ frontend/src/components/story/StoryViewer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef, useState } from "react";
 import { Chapter } from "../../types/story.types";
 import ReadingTimeBadge from "../ReadingTimeBadge";
 import toast from "react-hot-toast";
-import { AudioPlayer } from "../AudioPlayer"; // Sahi import path
+import AudioPlayer from "../AudioPlayer";
 
 interface Props {
   chapters: Chapter[];
